feat(applaunchpad): (backport) fix buggy port editing
