+++ frontend/providers/applaunchpad/src/pages/app/edit/components/Form.tsx
@@ -1,16 +1,11 @@
 import { obj2Query } from '@/api/tools';
 import MyIcon from '@/components/Icon';
 import { MyRangeSlider, MySelect, MySlider, MyTooltip, RangeInput, Tabs, Tip } from '@sealos/ui';
-import {
-  APPLICATION_PROTOCOLS,
-  defaultSliderKey,
-  defaultGpuSliderKey,
-  ProtocolList
-} from '@/constants/app';
+import { defaultSliderKey, defaultGpuSliderKey } from '@/constants/app';
 import { GpuAmountMarkList } from '@/constants/editApp';
 import { useToast } from '@/hooks/useToast';
 import { useGlobalStore } from '@/store/global';
-import { PVC_STORAGE_MAX, SEALOS_DOMAIN } from '@/store/static';
+import { PVC_STORAGE_MAX } from '@/store/static';
 import { useUserStore } from '@/store/user';
 import type { QueryType } from '@/types';
 import { type AppEditType } from '@/types/app';
@@ -31,33 +26,27 @@ import {
   Grid,
   IconButton,
   Input,
-  Switch,
-  Tooltip,
   useDisclosure,
   useTheme
 } from '@chakra-ui/react';
 import { throttle } from 'lodash';
-import { customAlphabet } from 'nanoid';
 import { useTranslation } from 'next-i18next';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useFieldArray, UseFormReturn } from 'react-hook-form';
 import type { ConfigMapType } from './ConfigmapModal';
-import type { CustomAccessModalParams } from './CustomAccessModal';
 import PriceBox from './PriceBox';
 import QuotaBox from './QuotaBox';
 import type { StoreType } from './StoreModal';
 import styles from './index.module.scss';
-import { mountPathToConfigMapKey, useCopyData } from '@/utils/tools';
+import { NetworkSection } from './NetworkSection';
+import { mountPathToConfigMapKey } from '@/utils/tools';
 import { useQuery } from '@tanstack/react-query';
 
-const CustomAccessModal = dynamic(() => import('./CustomAccessModal'));
 const ConfigmapModal = dynamic(() => import('./ConfigmapModal'));
 const StoreModal = dynamic(() => import('./StoreModal'));
 const EditEnvs = dynamic(() => import('./EditEnvs'));
-
-const nanoid = customAlphabet('abcdefghijklmnopqrstuvwxyz', 12);
 
 const labelWidth = 120;
 
@@ -69,7 +58,6 @@ const Form = ({
   pxVal,
   refresh,
   isAdvancedOpen,
-
   onDomainVerified
 }: {
   formHook: UseFormReturn<AppEditType, any>;
@@ -87,7 +75,6 @@ const Form = ({
   const { userSourcePrice } = useUserStore();
   const router = useRouter();
   const { toast } = useToast();
-  const { copyData } = useCopyData();
   const { name } = router.query as QueryType;
   const theme = useTheme();
   const isEdit = useMemo(() => !!name, [name]);
@@ -100,16 +87,6 @@ const Form = ({
     watch,
     formState: { errors }
   } = formHook;
-
-  const {
-    fields: networks,
-    append: appendNetworks,
-    remove: removeNetworks,
-    update: updateNetworks
-  } = useFieldArray({
-    control,
-    name: 'networks'
-  });
 
   const { fields: envs, replace: replaceEnvs } = useFieldArray({
     control,
@@ -169,7 +146,6 @@ const Form = ({
   );
 
   const [activeNav, setActiveNav] = useState(navList[0].id);
-  const [customAccessModalData, setCustomAccessModalData] = useState<CustomAccessModalParams>();
   const [configEdit, setConfigEdit] = useState<ConfigMapType>();
   const [storeEdit, setStoreEdit] = useState<StoreType>();
   const { isOpen: isEditEnvs, onOpen: onOpenEditEnvs, onClose: onCloseEditEnvs } = useDisclosure();
@@ -840,243 +816,12 @@ const Form = ({
             </Box>
           </Box>
 
-          {/* network */}
-          <Box id={'network'} {...boxStyles}>
-            <Box {...headerStyles}>
-              <MyIcon name={'network'} mr={'12px'} w={'24px'} color={'grayModern.900'} />
-              {t('Network Configuration')}
-            </Box>
-            <Box px={'42px'} py={'24px'} userSelect={'none'}>
-              {networks.map((network, i) => (
-                <Flex
-                  alignItems={'flex-start'}
-                  key={network.id}
-                  _notLast={{ pb: 6, borderBottom: theme.borders.base }}
-                  _notFirst={{ pt: 6 }}
-                >
-                  {/* Container Port Column - Fixed Width */}
-                  <Box w={'140px'}>
-                    <Box mb={'10px'} h={'20px'} fontSize={'base'} color={'grayModern.900'}>
-                      {t('Container Port')}
-                    </Box>
-                    <Input
-                      h={'32px'}
-                      type={'number'}
-                      w={'110px'}
-                      bg={'grayModern.50'}
-                      {...register(`networks.${i}.port`, {
-                        required:
-                          t('app.The container exposed port cannot be empty') ||
-                          'The container exposed port cannot be empty',
-                        valueAsNumber: true,
-                        min: {
-                          value: 1,
-                          message: t('app.The minimum exposed port is 1')
-                        },
-                        max: {
-                          value: 65535,
-                          message: t('app.The maximum number of exposed ports is 65535')
-                        }
-                      })}
-                    />
-                    {i === networks.length - 1 && networks.length + 1 <= 15 && (
-                      <Box mt={3}>
-                        <Button
-                          w={'100px'}
-                          variant={'outline'}
-                          leftIcon={<MyIcon name="plus" w={'18px'} fill={'#485264'} />}
-                          onClick={() =>
-                            appendNetworks({
-                              networkName: '',
-                              portName: nanoid(),
-                              port: 80,
-                              protocol: 'TCP',
-                              appProtocol: 'HTTP',
-                              openPublicDomain: false,
-                              publicDomain: '',
-                              customDomain: '',
-                              domain: SEALOS_DOMAIN,
-                              openNodePort: false,
-                              nodePort: undefined
-                            })
-                          }
-                        >
-                          {t('Add Port')}
-                        </Button>
-                      </Box>
-                    )}
-                  </Box>
-
-                  {/* Enable Internet Access Column - Fixed Width */}
-                  <Box w={'200px'} mx={7}>
-                    <Box mb={'8px'} h={'20px'} fontSize={'base'} color={'grayModern.900'}>
-                      {t('Open Public Access')}
-                    </Box>
-                    <Flex alignItems={'center'} h={'35px'}>
-                      <Switch
-                        className="driver-deploy-network-switch"
-                        size={'lg'}
-                        isChecked={!!network.openPublicDomain || !!network.openNodePort}
-                        onChange={(e) => {
-                          if (e.target.checked) {
-                            if (APPLICATION_PROTOCOLS.includes(network.appProtocol)) {
-                              updateNetworks(i, {
-                                ...getValues('networks')[i],
-                                networkName: network.networkName || `network-${nanoid()}`,
-                                protocol: 'TCP',
-                                appProtocol: network.appProtocol || 'HTTP',
-                                openPublicDomain: true,
-                                openNodePort: false,
-                                publicDomain: network.publicDomain || nanoid(),
-                                domain: network.domain || SEALOS_DOMAIN
-                              });
-                            } else {
-                              updateNetworks(i, {
-                                ...getValues('networks')[i],
-                                networkName: network.networkName || `network-${nanoid()}`,
-                                protocol: network.protocol,
-                                appProtocol: 'HTTP',
-                                openNodePort: true,
-                                openPublicDomain: false,
-                                customDomain: ''
-                              });
-                            }
-                          } else {
-                            updateNetworks(i, {
-                              ...getValues('networks')[i],
-                              openPublicDomain: false,
-                              openNodePort: false
-                            });
-                          }
-                        }}
-                      ></Switch>
-                    </Flex>
-                  </Box>
-
-                  {/* Protocol and Domain Column - Fixed Width */}
-                  <Box w={'500px'}>
-                    <Box mb={'8px'} h={'20px'}></Box>
-                    <Flex alignItems={'center'} h={'35px'}>
-                      {network.openPublicDomain || network.openNodePort ? (
-                        <>
-                          <MySelect
-                            width={'120px'}
-                            height={'32px'}
-                            borderTopRightRadius={0}
-                            borderBottomRightRadius={0}
-                            value={
-                              network.openPublicDomain
-                                ? network.appProtocol
-                                : network.openNodePort
-                                  ? network.protocol
-                                  : 'HTTP'
-                            }
-                            list={ProtocolList}
-                            onchange={(val: any) => {
-                              if (APPLICATION_PROTOCOLS.includes(val)) {
-                                updateNetworks(i, {
-                                  ...getValues('networks')[i],
-                                  protocol: 'TCP',
-                                  appProtocol: val,
-                                  openNodePort: false,
-                                  openPublicDomain: true,
-                                  networkName: network.networkName || `network-${nanoid()}`,
-                                  publicDomain: network.publicDomain || nanoid()
-                                });
-                              } else {
-                                updateNetworks(i, {
-                                  ...getValues('networks')[i],
-                                  protocol: val,
-                                  appProtocol: 'HTTP',
-                                  openNodePort: true,
-                                  openPublicDomain: false,
-                                  customDomain: ''
-                                });
-                              }
-                            }}
-                          />
-                          <Flex
-                            maxW={'350px'}
-                            flex={'1 0 0'}
-                            alignItems={'center'}
-                            h={'32px'}
-                            bg={'grayModern.50'}
-                            px={4}
-                            border={theme.borders.base}
-                            borderLeft={0}
-                            borderTopRightRadius={'md'}
-                            borderBottomRightRadius={'md'}
-                          >
-                            <Tooltip label={t('click_to_copy_tooltip')}>
-                              <Box
-                                flex={1}
-                                userSelect={'all'}
-                                className="textEllipsis"
-                                onClick={() => {
-                                  copyData(`${network.publicDomain}.${network.domain}`);
-                                }}
-                              >
-                                {network.customDomain
-                                  ? network.customDomain
-                                  : network.openNodePort
-                                    ? network?.nodePort
-                                      ? `${network.protocol.toLowerCase()}.${network.domain}:${
-                                          network.nodePort
-                                        }`
-                                      : `${network.protocol.toLowerCase()}.${network.domain}:${t(
-                                          'pending_to_allocated'
-                                        )}`
-                                    : `${network.publicDomain}.${network.domain}`}
-                              </Box>
-                            </Tooltip>
-
-                            {network.openPublicDomain && !network.openNodePort && (
-                              <Box
-                                fontSize={'11px'}
-                                color={'brightBlue.600'}
-                                cursor={'pointer'}
-                                onClick={() =>
-                                  setCustomAccessModalData({
-                                    publicDomain: network.publicDomain,
-                                    currentCustomDomain: network.customDomain,
-                                    domain: network.domain
-                                  })
-                                }
-                              >
-                                {t('Custom Domain')}
-                              </Box>
-                            )}
-                          </Flex>
-                        </>
-                      ) : (
-                        <Box w={'470px'} h={'32px'}></Box>
-                      )}
-                    </Flex>
-                  </Box>
-
-                  {/* Delete Button Column - Fixed Width */}
-                  {networks.length > 1 && (
-                    <Box w={'50px'} ml={3}>
-                      <Box mb={'8px'} h={'20px'}></Box>
-                      <IconButton
-                        height={'32px'}
-                        width={'32px'}
-                        aria-label={'button'}
-                        variant={'outline'}
-                        bg={'#FFF'}
-                        _hover={{
-                          color: 'red.600',
-                          bg: 'rgba(17, 24, 36, 0.05)'
-                        }}
-                        icon={<MyIcon name={'delete'} w={'16px'} fill={'#485264'} />}
-                        onClick={() => removeNetworks(i)}
-                      />
-                    </Box>
-                  )}
-                </Flex>
-              ))}
-            </Box>
-          </Box>
+          <NetworkSection
+            formHook={formHook}
+            onDomainVerified={onDomainVerified}
+            boxStyles={boxStyles}
+            headerStyles={headerStyles}
+          />
           {/* settings */}
           {already && (
             <Accordion
@@ -1368,23 +1113,6 @@ const Form = ({
           )}
         </Box>
       </Grid>
-      {!!customAccessModalData && (
-        <CustomAccessModal
-          {...customAccessModalData}
-          onClose={() => setCustomAccessModalData(undefined)}
-          onSuccess={(e) => {
-            const i = networks.findIndex(
-              (item) => item.publicDomain === customAccessModalData.publicDomain
-            );
-            if (i === -1) return;
-            updateNetworks(i, {
-              ...networks[i],
-              customDomain: e
-            });
-            onDomainVerified?.({ index: i, customDomain: e });
-          }}
-        />
-      )}
       {isEditEnvs && (
         <EditEnvs defaultEnv={envs} onClose={onCloseEditEnvs} successCb={(e) => replaceEnvs(e)} />
       )}
